test(server): avoid an objectstore test dependency cycle
