+++ crates/loonfs-objectstore/Cargo.toml
@@ -30,7 +30,6 @@ thiserror.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-loonfs-test-support.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = ["test-util"] }
 toml.workspace = true
